gestion-alternants/backend/requirements.txt
Django>=4.2.5
psycopg2-binary>=2.9
djangorestframework>=3.14.0
djangorestframework-simplejwt
django-cors-headers
drf_yasg